src/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
python-decouple==3.8
pytils==0.4.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
